DynamicEnum, ClassType: moved from utils to base
