+++ base/enum.py
@@ -5,7 +5,7 @@ from typing import Optional, Callable, Iterable, Union, Type
 try:  # Assume we're a sub-module in a package.
     from utils import arguments as arg
 except ImportError:  # Apparently no higher-level package has been imported, fall back to a local import.
-    from . import arguments as arg
+    from ..utils import arguments as arg
 
 Name = str
 Value = Union[str, int, arg.Auto]
