+++ connectors/conn_type.py
@@ -1,7 +1,7 @@
 try:  # Assume we're a sub-module in a package.
-    from utils.enum import ClassType
+    from base.enum import ClassType
 except ImportError:  # Apparently no higher-level package has been imported, fall back to a local import.
-    from ..utils.enum import ClassType
+    from ..base.enum import ClassType
 
 
 class ConnType(ClassType):
