+++ streams/stream_type.py
@@ -6,10 +6,10 @@ TMP_FILES_ENCODING = 'utf8'
 
 try:  # Assume we're a sub-module in a package.
     from utils import arguments as arg
-    from utils.enum import ClassType
+    from base.enum import ClassType
 except ImportError:  # Apparently no higher-level package has been imported, fall back to a local import.
     from ..utils import arguments as arg
-    from ..utils.enum import ClassType
+    from ..base.enum import ClassType
 
 DICT_METHOD_SUFFIX = dict(
     AnyStream='any_stream',
